"crispy-forms wurde hinzugefügt."

+++ blog/views.py
@@ -1,4 +1,5 @@
 from multiprocessing import context
+from django.http import HttpResponse
 from django.shortcuts import get_object_or_404, render, redirect
 from .models import Post, Like
 from .forms import CommentForm, PostForm
@@ -45,6 +46,8 @@ def post_detail(request, slug):
 def post_update(request, slug):
     obj= get_object_or_404(Post, slug=slug)
     form = PostForm(request.POST or None, request.FILES or None, instance=obj)
+    if request.user.id != obj.author.id:
+        return HttpResponse("You're not authorized!!")
     if form.is_valid():
         form.save()
         return redirect("blog:list")
@@ -56,6 +59,8 @@ def post_update(request, slug):
 
 def post_delete(request, slug):
     obj= get_object_or_404(Post, slug=slug)
+    if request.user.id != obj.author.id:
+        return HttpResponse("You're not authorized!!")
     if request.method=='POST':
         obj.delete()
         return redirect("blog:list")
